Ammo Pack resupply changes
a weapon with no initial reserve ammo will now be granted 1/4 of its total ammo when picking up an Ammo Pack.
(can't seem to access the weapon class from here to get a fraction of MagAmmo, and trying to grab the amount given from an ammo pickup didn't work either, but if there is a workaround feel free to change this. the difficulty comes from the Ammo Pack resupplying vanilla weapons, thus using base UT classes for its variables)