Noots Fixes
Moved and rotated the ICIS third person mesh so it makes a bit more sense in third person as before it just was held up instead of down as its in first person.
Fixed missing Ammoclass error for the CYLO and CYLOFirestorm as the firestorm or CYLO could not load the correct ammoclass this is caused when 2 weapons that extend each other are in the loadout as the same that is when this error occurs in this case the firestorm used the CYLO as a base.
Removed Pistol anims for the CYLO, why the balls is it even there for a rifle :p.
Fixed CYLOFirestorm third person mesh missing a brass bone reference (Bound the Ejector to the Root bone as the third person mesh lacks a proper ejector bone, its not accurate but hey whatever works)
Fixed CYLO third person mesh missing a brass bone reference (Bound the Ejector to the Root bone as the third person mesh lacks a proper ejector bone, its not accurate but hey whatever works)
Fixed missing Ammoclass error for the M46 and M46 RDS
Moved and rotated the X8 knife third person mesh so it makes a bit more sense in third person as before it just was held up instead of down as its in first person.
Fixed MARS third person mesh missing a brass and tip bone reference.
Fixed PD-97 third person mesh missing a brass bone reference (Bound the Ejector to the Base bone as the third person mesh lacks a proper ejector bone, its not accurate but hey whatever works)
Fixed CYLO having wrong textures applied to its third person model.
Fixed Mars-2 and Mars-3 both having the wrong grenadefire anim assigned as it seems the name got changed but was not updated in the code.
Added sightfire and prefire code and anim references to both Mars-2 and Mars-3
Fixed a typo in the animation sequence for FPA_DTS where bLock was supposed to be BlockIdle this also fixes an error where the BlockIdle animation could not be found for the DragonTooth.
Fixed AH208 missing animation error by using the same code as the RS8 as the AH208 had code where when scoped it should use a different fireanim= animation but there was no sequence to pick from so but somehow the ADS anim worked fine in-game however I simply copied over the RS8 ADSanim code and it fixed the error
Fixed MAU-52 missing fire animation error by assigning the Point animation to it instead but sadly due to how the scope etc works u will never see the animation play.
Fixed CX85 third person mesh missing a ejector bone reference (Bound the Ejector to the Weapon bone as the third person mesh lacks a proper ejector bone, its not accurate but hey whatever works)
Removed mesh location code for ARC.
Fixed R9 team skin issue
Fixed XOXO third person mesh missing a tip bone reference.
(Asset Files have changed but wont affect compile due to bone updates on models, Just use old asset files for now and will send you new ones later)