Added option to turn off the damage view flash
(This was a 2 min job so if it can be done better let me know)

Added a toggle to turn off the viewflash effects added since 1.4.

If a server has the viewflash enabled it will be enforced for all clients so a player cannot turn it off on a server that has it enabled.

When a player plays a offline match or host there own server they will be able to toggle the viewflash on or off as they see fit