Small fix for screen hitching when shot by high RPM weapons
Fixed a issue where the screen would hitch when u got hit by a high RPM weapon.

Weirdly enough the feature that is in the code to prevent this *rolling view* does work but not for high RPM weapons, setting the min damage value to 0 fixes the weird hitching when u get hit by high RPM guns.