Crosshairs & WUI
Crosshairs now work for melee weapons, shotguns & A500 Reptile. Meaning the crosshair code is complete for use. All that needs to be added are the defaults to the weapons.
WUI now scaled down to contain the mass amount of weapons in the mod. I have plans for this later on