(Experimental) eyeheight fix from utcomp 1.8b
This fixes a annoying bug where in online matches your view sinks into the stairs, terrain or sloped surfaces.
I have tried this fix in the past but for some it broke footstep audio whilst for other it didn't. We will need to test this change pretty hard to see if anything breaks but i already did some offline testing and found no issues